Decompile DRA func_80103ED4
