+++ config/symbols.us.dra.txt
@@ -143,7 +143,6 @@ D_800DBFFC = 0x800DBFFC; // ignore:true
 aXXVQ = 0x800DCBF4;
 aAxearmor = 0x800DCC00;
 jpt_80102918 = 0x800DCC2C;
-jpt_80103F08 = 0x800DCDC0;
 aGold = 0x800DD0A8;
 aNz0 = 0x800DD0B0;
 aFNz0 = 0x800DD0B4;
@@ -185,14 +184,6 @@ aAtariNuki = 0x800E1AA4;
 aLightTimer02x = 0x800E1E48;
 SetGameState = 0x800E4124;
 HandleGameOver = 0x800E5584;
-MemcardInit = 0x800E92E4;
-MemcardParse = 0x800E930C;
-GetMemcardFreeBlockCount = 0x800E9508;
-MemcardDetectSave = 0x800E9530;
-MemcardWriteFile = 0x800E96E8;
-MemcardClose = 0x800E9804;
-StoreSaveData = 0x800E9C14;
-MakeMemcardPath = 0x800EA48C;
 DecompressWriteNibble = 0x800EAFC8;
 DecompressReadNibble = 0x800EB03C;
 DecompressData = 0x800EB098;
