+++ src/dra/62D70.c
@@ -493,239 +493,3 @@ void func_80103EAC(void) {
     D_80137E4C = 0;
     MemcardInfoInit();
 }
-
-INCLUDE_ASM("dra/nonmatchings/62D70", func_80103ED4);
-
-void func_8010427C(void) {
-    FreePrimitives(D_80137E40);
-    FreePrimitives(D_80137E44);
-    FreePrimitives(D_80137E48);
-}
-
-INCLUDE_ASM("dra/nonmatchings/62D70", func_801042C4);
-
-INCLUDE_ASM("dra/nonmatchings/62D70", func_80104790);
-
-INCLUDE_ASM("dra/nonmatchings/62D70", func_80105078);
-
-void func_80105408(void) {
-    g_Player.D_80072EF4 = 0x1000;
-    g_Player.D_80072EFC = 1;
-}
-
-INCLUDE_ASM("dra/nonmatchings/62D70", func_80105428);
-
-void DestroyEntity(Entity* entity) {
-    s32 i;
-    s32 length;
-    u32* ptr;
-
-    if (entity->flags & FLAG_HAS_PRIMS) {
-        FreePrimitives(entity->primIndex);
-    }
-
-    ptr = (u32*)entity;
-    length = sizeof(Entity) / sizeof(u32);
-    for (i = 0; i < length; i++)
-        *ptr++ = NULL;
-}
-
-void DestroyEntities(s16 startIndex) {
-    Entity* pItem;
-
-    for (pItem = &g_Entities[startIndex];
-         pItem < &g_Entities[TOTAL_ENTITY_COUNT]; pItem++)
-        DestroyEntity(pItem);
-}
-
-void DrawEntitiesHitbox(s32 blendMode) {
-    DR_MODE* drawMode;
-    s32 polyCount;
-    s32* ot;
-    Entity* entity;
-    TILE* tile;
-    u32 otIdx;
-    u16 x;
-    u16 y;
-
-    ot = g_CurrentBuffer->ot;
-    tile = &g_CurrentBuffer->tiles[g_GpuUsage.tile];
-    drawMode = &g_CurrentBuffer->drawModes[g_GpuUsage.drawModes];
-    otIdx = 0x1F0;
-    for (polyCount = 0, entity = g_Entities; polyCount < 0x40; polyCount++,
-        entity++) {
-        if (entity->hitboxState == 0)
-            continue;
-        if (g_GpuUsage.tile >= GPU_MAX_TILE_COUNT) {
-            break;
-        }
-
-        y = (u16)entity->posY.i.hi + (u16)g_backbufferY;
-        x = (u16)entity->posX.i.hi + (u16)g_backbufferX;
-        if (entity->facingLeft) {
-            x -= entity->hitboxOffX;
-        } else {
-            x += entity->hitboxOffX;
-        }
-        y += entity->hitboxOffY;
-
-        tile->r0 = 0xFF;
-        tile->g0 = 0xFF;
-        tile->b0 = 0xFF;
-        if (entity->hitboxState == 2) {
-            tile->r0 = 0;
-            tile->g0 = 0xFF;
-            tile->b0 = 0;
-        }
-        tile->x0 = x - entity->hitboxWidth;
-        tile->y0 = y - entity->hitboxHeight;
-        tile->w = entity->hitboxWidth * 2;
-        tile->h = entity->hitboxHeight * 2;
-        SetSemiTrans(tile, 1);
-        AddPrim(&ot[otIdx], tile);
-        tile++;
-        g_GpuUsage.tile++;
-    }
-
-    for (; polyCount < GPU_MAX_TILE_COUNT; polyCount++, entity++) {
-        if (entity->hitboxState == 0)
-            continue;
-        if (g_GpuUsage.tile >= GPU_MAX_TILE_COUNT) {
-            break;
-        }
-
-        y = (u16)entity->posY.i.hi + (u16)g_backbufferY;
-        x = (u16)entity->posX.i.hi + (u16)g_backbufferX;
-        if (entity->facingLeft) {
-            x -= entity->hitboxOffX;
-        } else {
-            x += entity->hitboxOffX;
-        }
-        y += entity->hitboxOffY;
-
-        tile->r0 = 0xFF;
-        tile->g0 = 0xFF;
-        tile->b0 = 0xFF;
-        if (entity->hitboxState == 1) {
-            tile->r0 = 0xFF;
-            tile->g0 = 0;
-            tile->b0 = 0;
-        }
-        if (entity->hitboxState == 2) {
-            tile->r0 = 0;
-            tile->g0 = 0;
-            tile->b0 = 0xFF;
-        }
-        if (entity->hitboxState == 3) {
-            tile->r0 = 0xFF;
-            tile->g0 = 0;
-            tile->b0 = 0xFF;
-        }
-        tile->x0 = x - entity->hitboxWidth;
-        tile->y0 = y - entity->hitboxHeight;
-        tile->w = entity->hitboxWidth * 2;
-        tile->h = entity->hitboxHeight * 2;
-        SetSemiTrans(tile, 1);
-        AddPrim(&ot[otIdx], tile);
-        tile++;
-        g_GpuUsage.tile++;
-    }
-
-    if (g_GpuUsage.drawModes < MAX_DRAW_MODES) {
-        SetDrawMode(drawMode, 0, 0, (blendMode - 1) << 5, &g_Vram.D_800ACD80);
-        AddPrim(&ot[otIdx], drawMode);
-        g_GpuUsage.drawModes++;
-    }
-}
-
-INCLUDE_ASM("dra/nonmatchings/62D70", func_80106A28);
-
-bool LoadMonsterLibrarianPreview(s32 monsterId) {
-    if (g_IsUsingCd)
-        return false;
-
-    if (!g_UseDisk) {
-        if (LoadFileSim(monsterId, SimFileType_Monster) < 0) {
-            return false;
-        }
-    } else {
-        g_CdStep = CdStep_LoadInit;
-        g_LoadFile = CdFile_Monster;
-        g_LoadOvlIdx = monsterId;
-    }
-    return true;
-}
-
-void func_801071CC(POLY_GT4* poly, u32 colorIntensity, s32 vertexIndex) {
-    switch (vertexIndex) {
-    case 0:
-        poly->b0 = colorIntensity;
-        poly->g0 = colorIntensity;
-        poly->r0 = colorIntensity;
-        break;
-    case 1:
-        poly->b1 = colorIntensity;
-        poly->g1 = colorIntensity;
-        poly->r1 = colorIntensity;
-        break;
-    case 2:
-        poly->b2 = colorIntensity;
-        poly->g2 = colorIntensity;
-        poly->r2 = colorIntensity;
-        break;
-    case 3:
-        poly->b3 = colorIntensity;
-        poly->g3 = colorIntensity;
-        poly->r3 = colorIntensity;
-        break;
-    }
-}
-
-void func_80107250(POLY_GT4* poly, s32 colorIntensity) {
-    func_801071CC(poly, (u8)colorIntensity, 0);
-    func_801071CC(poly, (u8)colorIntensity, 1);
-    func_801071CC(poly, (u8)colorIntensity, 2);
-    func_801071CC(poly, (u8)colorIntensity, 3);
-}
-
-void func_801072BC(POLY_GT4* poly) { func_80107250(poly, 0); }
-
-void func_801072DC(POLY_GT4* poly) { func_80107250(poly, 0x80); }
-
-void func_801072FC(POLY_G4* poly) {
-    setRGB0(poly, 0, 0, 0);
-    setRGB1(poly, 0, 0, 0);
-    setRGB2(poly, 0, 0, 0);
-    setRGB3(poly, 0, 0, 0);
-}
-
-void SetPrimRect(Primitive* poly, s32 x, s32 y, s32 width, s32 height) {
-    poly->x0 = x;
-    poly->y0 = y;
-    poly->x1 = x + width;
-    poly->y1 = y;
-    poly->x2 = x;
-    poly->x3 = x + width;
-    poly->y2 = y + height;
-    poly->y3 = y + height;
-}
-
-void SetTexturedPrimRect(
-    Primitive* poly, s32 x, s32 y, s32 width, s32 height, s32 u, s32 v) {
-    poly->x0 = x;
-    poly->y0 = y;
-    poly->x1 = x + width;
-    poly->y1 = y;
-    poly->x2 = x;
-    poly->y2 = y + height;
-    poly->x3 = x + width;
-    poly->y3 = y + height;
-    poly->v0 = v;
-    poly->v1 = v;
-    poly->u0 = u;
-    poly->u1 = u + width;
-    poly->u2 = u;
-    poly->v2 = v + height;
-    poly->u3 = u + width;
-    poly->v3 = v + height;
-}
